% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{makePlotsParticleMetropolisHastingsSVModel}
\alias{makePlotsParticleMetropolisHastingsSVModel}
\title{Make plots for tutorial}
\usage{
makePlotsParticleMetropolisHastingsSVModel(y, res, noBurnInIterations,
  noIterations, nPlot)
}
\arguments{
\item{y}{Observations from the model for \eqn{t=1,...,T}.}

\item{res}{The output from a run of particleMetropolisHastings,
particleMetropolisHastingsSVmodel or
particleMetropolisHastingsSVmodelReparameterised.}

\item{noBurnInIterations}{The number of burn-in iterations in the 
PMH algorithm. Must be smaller than noIterations.}

\item{noIterations}{The number of iterations in the PMH algorithm.}

\item{nPlot}{Number of steps in the Markov chain to plot.}
}
\value{
The function returns plots similar to the ones in the reference as
well as the estimate of the integrated autocorrelation time
for each parameter.
}
\description{
Creates diagnoistic plots from runs of the particle Metropolis-Hastings algorithm.
}
\references{
Dahlin, J. & Schon, T. B. "Getting started with particle 
Metropolis-Hastings for inference in nonlinear dynamical models." 
pre-print, arXiv:1511.01707, 2017.
}
\author{
Johan Dahlin <uni (at) johandahlin.com.nospam>s
}
